requirements.txt
mysql-connector-python==8.0.20
pandas==1.0.4
psycopg2-binary==2.8.5